+++ src/main/java/com/moru/backend/domain/auth/api/AuthController.java
@@ -1,6 +1,6 @@
 package com.moru.backend.domain.auth.api;
 
-import com.moru.backend.domain.auth.application.AuthService;
+import com.moru.backend.domain.auth.application.*;
 import com.moru.backend.domain.auth.dto.*;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.servlet.http.HttpServletRequest;
@@ -16,31 +16,34 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/auth")
 @RequiredArgsConstructor
 public class AuthController {
-    private final AuthService authService;
+    private final SignupService signupService;
+    private final LoginService loginService;
+    private final RefreshService refreshService;
+    private final LogoutService logoutService;
 
     @Operation(summary = "회원가입")
     @PostMapping("/signup")
     public ResponseEntity<Void> signup(@RequestBody @Valid SignupRequest request) {
-        authService.signup(request);
+        signupService.signup(request);
         return ResponseEntity.ok().build();
     }
 
     @Operation(summary = "로그인")
     @PostMapping("/login")
     public ResponseEntity<TokenResponse> login(@RequestBody @Valid LoginRequest request) {
-        return ResponseEntity.ok(authService.login(request));
+        return ResponseEntity.ok(loginService.login(request));
     }
 
     @Operation(summary = "토큰 재발급 (access & refresh 모두 새로 만들어짐)")
     @PostMapping("/refresh")
     public ResponseEntity<TokenResponse> refresh(@RequestBody @Valid TokenRefreshRequest request) {
-        return ResponseEntity.ok(authService.refreshToken(request));
+        return ResponseEntity.ok(refreshService.refreshToken(request));
     }
 
     @Operation(summary = "로그아웃")
     @PostMapping("/logout")
     public ResponseEntity<Void> logout(HttpServletRequest request) {
-        authService.logout(request);
+        logoutService.logout(request);
         return ResponseEntity.ok().build();
     }
 }
